修改了通信协议为艾洛克智能锁协议，增加了重发无线机制

+++ 智能门锁MCU固件程序/STM32L031K6Tx/Src/uart.c
@@ -41,7 +41,7 @@ void UartSendData(USART_TypeDef *UARTx, uint8_t byte)
 }
 void UartSendBytes(USART_TypeDef *UARTx, uint8_t *buf, uint16_t len)
 {
-    uint8_t i = 0;
+    uint16_t i = 0;
     for (i = 0; i < len; i++)
     {
         UartSendData(UARTx,*buf++);
